requirements.txt
numpy
torch
matplotlib
pydantic
tqdm
imageio
Pillow
tensorboard
scipy